// src/app/interfaces/cron-job-util.ts
export interface CronJobUtil {
    timeUnit: string;
    dayOfWeek: string;
    dayOfMonth: string;
    time: string;
}

export const Times = ["יום", "שבוע", "חודש"];
export const TimesTo = ["בשעה", "ביום", "ב"];
export const Days = ["ראשון", "שני", "שלישי", "רביעי", "חמישי", "שישי", "שבת"];